Scratch buffer copy-based implementation of ncclBcast / ncclBroadcast (#417)
Encountered a hang with rccl-test's ncclBcast runs. In rccl-test with
ncclBcast, the buffer changes only for the root-gpu b/w test runs. So,
the channel key changes only for the root gpu and remains the same for
the other gpus. So, the root gpu starts to create a new sm channel while
the other gpus do not create new sm channels. Hence, a hang occurs.

To fix this, we have to use copy-based implementation. This is the
simplest copy-based implementation. Uses a static channel key with sm
channels created using scratch buffer as the remote memories.