DeviceSemaphore  fix (#553)
Fix the bug, make sure a thread will be wake up if semaphore be
released.

This pull request includes a modification to the `DeviceSemaphore`
struct in the `concurrency_device.hpp` file, specifically in the
`acquire` method. The change refines the logic for acquiring a semaphore
by adjusting the condition used to handle contention scenarios.